GCA-30 Update Gym services - added dtos for user and refactoring other dtos
