+++ src/main/java/com/gca/dto/trainer/TrainerCreateRequest.java
@@ -2,12 +2,12 @@ package com.gca.dto.trainer;
 
 import com.gca.model.TrainingType;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
 
 @Data
-@SuperBuilder(toBuilder = true)
+@Builder(toBuilder = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class TrainerCreateRequest {
