+++ src/main/java/com/gca/dto/training/TrainingCreateRequest.java
@@ -1,14 +1,14 @@
 package com.gca.dto.training;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
 
 import java.time.LocalDate;
 
 @Data
-@SuperBuilder(toBuilder = true)
+@Builder(toBuilder = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class TrainingCreateRequest {
